backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bitrix24-rest==1.1.1
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-pytest==0.2.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.4
exceptiongroup==1.0.4
flake8==6.0.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
outcome==1.2.0
packaging==22.0
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PySocks==1.7.1
pytest==7.2.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.8.5
wsproto==1.2.0
